[core][rdt] Initial refactor for bring your own transport
