+++ python/ray/_private/serialization.py
@@ -722,7 +722,7 @@ class SerializationContext:
             obj_id is not None
         ), "`obj_id` is required, and it is the key to retrieve corresponding tensors from the GPU object store."
         # Regardless of whether `tensors` is empty, we always store the GPU object
-        # in the GPU object store. This ensures that `get_tensor_transport_metadata` is not
+        # in the GPU object store. This ensures that `__ray_get_tensor_transport_metadata__` is not
         # blocked indefinitely.
         worker = ray._private.worker.global_worker
         gpu_object_manager = worker.gpu_object_manager
