+++ python/ray/experimental/collective/__init__.py
@@ -9,7 +9,6 @@ from ray.experimental.collective.operations import (
     allreduce,
     reducescatter,
 )
-from ray.experimental.collective.util import get_tensor_transport_manager
 
 __all__ = [
     "allgather",
@@ -19,5 +18,4 @@ __all__ = [
     "create_collective_group",
     "destroy_collective_group",
     "destroy_all_collective_groups",
-    "get_tensor_transport_manager",
 ]
