+++ python/ray/experimental/gpu_object_manager/gpu_object_manager.py
@@ -9,6 +9,7 @@ import ray
 from ray._private import ray_constants
 from ray._private.custom_types import TensorTransportEnum
 from ray._raylet import ObjectRef
+from ray.util.annotations import PublicAPI
 
 if TYPE_CHECKING:
     import torch
@@ -16,7 +17,10 @@ if TYPE_CHECKING:
     from ray.experimental.gpu_object_manager.gpu_object_store import (
         GPUObjectStore,
     )
-    from ray.util.collective.types import CommunicatorMetadata, TensorTransportMetadata
+    from ray.experimental.gpu_object_manager.types import (
+        CommunicatorMetadata,
+        TensorTransportMetadata,
+    )
 
 logger = logging.getLogger(__name__)
 
@@ -28,8 +32,6 @@ logger = logging.getLogger(__name__)
 #   of a tensor in the GPU object store.
 class GPUObjectMeta(NamedTuple):
     src_actor: "ray.actor.ActorHandle"
-    # Must be a valid backend name as defined in
-    # `ray.util.collective.types.Backend`.
     tensor_transport_backend: str
     tensor_transport_meta: "TensorTransportMetadata"
     # sent_dest_actors tracks the set of actor IDs that this object has been sent to.
@@ -51,8 +53,7 @@ class TransferMetadata(NamedTuple):
     timeout: float
 
 
-# TODO(swang): Uncomment and add an API docs page and example usage.
-# @PublicAPI(stability="alpha")
+@PublicAPI(stability="alpha")
 def wait_tensor_freed(tensor: "torch.Tensor", timeout: Optional[float] = None):
     """
     Wait for the tensor to be freed.
@@ -181,14 +182,16 @@ class GPUObjectManager:
         Cleans up the ref_info_map, kill the src and dst actors, and destroy the
         collective group if necessary.
         """
-        from ray.experimental.collective import (
-            destroy_collective_group,
-            get_tensor_transport_manager,
+        from ray.experimental.collective import destroy_collective_group
+        from ray.experimental.gpu_object_manager.collective_tensor_transport import (
+            CollectiveCommunicatorMetadata,
         )
         from ray.experimental.gpu_object_manager.gpu_object_store import (
             __ray_abort_transport__,
         )
-        from ray.util.collective.types import CollectiveCommunicatorMetadata
+        from ray.experimental.gpu_object_manager.util import (
+            get_tensor_transport_manager,
+        )
 
         ref_info = ref_info_map.pop(failed_ref.hex(), None)
         if ref_info is None:
@@ -292,8 +295,8 @@ class GPUObjectManager:
             tensor_transport: The tensor transport protocol to use for the GPU object.
             tensor_transport_meta: The tensor transport metadata that is pre-computed.
         """
-        from ray.experimental.collective import get_tensor_transport_manager
         from ray.experimental.gpu_object_manager.gpu_object_store import (
+            __ray_get_tensor_transport_metadata__,
             _tensor_transport_to_collective_backend,
         )
 
@@ -301,12 +304,17 @@ class GPUObjectManager:
             tensor_transport
         )
         obj_id = obj_ref.hex()
-        tensor_transport_manager = get_tensor_transport_manager(
-            tensor_transport_backend
-        )
         if not tensor_transport_meta:
-            tensor_meta = tensor_transport_manager.get_tensor_transport_metadata(
-                src_actor, obj_id
+            # Submit a Ray actor task to the source actor to get the tensor metadata.
+            # The metadata is a list of tuples, where each tuple contains the shape and dtype
+            # of a tensor in the GPU object store. This function returns an ObjectRef that
+            # points to the tensor metadata.
+            # NOTE(swang): We put this task on the background thread to avoid tasks
+            # executing on the main thread blocking this task.
+            tensor_meta = src_actor.__ray_call__.options(
+                concurrency_group="_ray_system"
+            ).remote(
+                __ray_get_tensor_transport_metadata__, obj_id, tensor_transport_backend
             )
         else:
             tensor_meta = tensor_transport_meta
@@ -342,9 +350,11 @@ class GPUObjectManager:
         Returns:
             None
         """
-        from ray.experimental.collective import get_tensor_transport_manager
         from ray.experimental.gpu_object_manager.gpu_object_store import (
             __ray_fetch_gpu_object__,
+        )
+        from ray.experimental.gpu_object_manager.util import (
+            get_tensor_transport_manager,
         )
 
         if tensor_transport not in [
@@ -425,10 +435,12 @@ class GPUObjectManager:
             if self.is_managed_object(arg.hex()):
                 gpu_object_refs.add(arg)
         if gpu_object_refs:
-            from ray.experimental.collective import get_tensor_transport_manager
             from ray.experimental.gpu_object_manager.gpu_object_store import (
                 __ray_recv__,
                 __ray_send__,
+            )
+            from ray.experimental.gpu_object_manager.util import (
+                get_tensor_transport_manager,
             )
 
         # Count the number of readers for each GPU object.
@@ -604,9 +616,11 @@ class GPUObjectManager:
         """
         # Import get_collective_groups here to avoid dependency on
         # collective libraries for default Ray installation.
-        from ray.experimental.collective import get_tensor_transport_manager
         from ray.experimental.gpu_object_manager.gpu_object_store import (
             _tensor_transport_to_collective_backend,
+        )
+        from ray.experimental.gpu_object_manager.util import (
+            get_tensor_transport_manager,
         )
 
         tensor_transport_backend = _tensor_transport_to_collective_backend(
@@ -632,9 +646,11 @@ class GPUObjectManager:
             tensors: The tensors to put into the GPU object manager.
 
         """
-        from ray.experimental.collective import get_tensor_transport_manager
         from ray.experimental.gpu_object_manager.gpu_object_store import (
             _tensor_transport_to_collective_backend,
+        )
+        from ray.experimental.gpu_object_manager.util import (
+            get_tensor_transport_manager,
         )
 
         tensor_transport_backend = _tensor_transport_to_collective_backend(
