+++ python/ray/experimental/gpu_object_manager/gpu_object_store.py
@@ -7,12 +7,13 @@ import ray
 import ray.util.collective as collective
 from ray._private.custom_types import TensorTransportEnum
 from ray._raylet import ObjectRef
-from ray.experimental.collective import get_tensor_transport_manager
-from ray.experimental.collective.util import device_match_transport
-from ray.util.collective.types import (
-    Backend,
+from ray.experimental.gpu_object_manager.types import (
     CommunicatorMetadata,
     TensorTransportMetadata,
+)
+from ray.experimental.gpu_object_manager.util import (
+    device_match_transport,
+    get_tensor_transport_manager,
 )
 
 try:
@@ -24,21 +25,39 @@ except ImportError:
     )
 
 TENSOR_TRANSPORT_TO_COLLECTIVE_BACKEND = {
-    TensorTransportEnum.NCCL: Backend.NCCL,
-    TensorTransportEnum.GLOO: Backend.TORCH_GLOO,
-    TensorTransportEnum.NIXL: Backend.NIXL,
+    TensorTransportEnum.NCCL: "nccl",
+    TensorTransportEnum.GLOO: "torch_gloo",
+    TensorTransportEnum.NIXL: "nixl",
 }
 
 
 def _tensor_transport_to_collective_backend(
     tensor_transport: TensorTransportEnum,
-) -> Backend:
+) -> str:
     try:
         return TENSOR_TRANSPORT_TO_COLLECTIVE_BACKEND[tensor_transport]
     except KeyError:
         raise ValueError(
             f"Invalid tensor transport {tensor_transport.name}, must be one of {list(TENSOR_TRANSPORT_TO_COLLECTIVE_BACKEND.keys())}."
         )
+
+
+def __ray_get_tensor_transport_metadata__(
+    self, obj_id: str, backend: str
+) -> TensorTransportMetadata:
+    """Helper function that runs on the src actor to get transport metadata."""
+    from ray._private.worker import global_worker
+
+    gpu_object_store = global_worker.gpu_object_manager.gpu_object_store
+    # NOTE: We do not specify a timeout here because the user task that returns
+    # it could take arbitrarily long and we don't want to trigger a spurious
+    # timeout.
+    gpu_object = gpu_object_store.wait_and_get_object(obj_id)
+
+    tensor_transport_manager = get_tensor_transport_manager(backend)
+    return tensor_transport_manager.extract_tensor_transport_metadata(
+        obj_id, gpu_object
+    )
 
 
 def __ray_send__(
@@ -129,12 +148,11 @@ def __ray_abort_transport__(self, obj_id: str, communicator_meta: CommunicatorMe
 def __ray_free__(
     self,
     obj_id: str,
-    tensor_transport_backend: Backend,
+    tensor_transport_backend: str,
     tensor_transport_meta: TensorTransportMetadata,
 ):
     try:
         from ray._private.worker import global_worker
-        from ray.experimental.collective import get_tensor_transport_manager
 
         tensor_transport_manager = get_tensor_transport_manager(
             tensor_transport_backend
