+++ python/ray/util/collective/types.py
@@ -3,16 +3,14 @@
 from dataclasses import dataclass
 from datetime import timedelta
 from enum import Enum
-from typing import TYPE_CHECKING, Any, List, Optional, Tuple
-
-from numpy import int32
+from typing import TYPE_CHECKING
 
 _NUMPY_AVAILABLE = True
 _TORCH_AVAILABLE = True
 _CUPY_AVAILABLE = True
 
 if TYPE_CHECKING:
-    import torch
+    pass
 
 try:
     import torch as th  # noqa: F401
@@ -57,71 +55,6 @@ class Backend(object):
         return backend
 
 
-@dataclass
-class TensorTransportMetadata:
-    """Metadata for tensors stored in the GPU object store.
-
-    Args:
-        tensor_meta: A list of tuples, each containing the shape and dtype of a tensor.
-        tensor_device: The device of the tensor. Currently, we require all tensors in the
-        list have the same device type.
-    """
-
-    tensor_meta: List[Tuple["torch.Size", "torch.dtype"]]
-    tensor_device: Optional["torch.device"] = None
-
-
-@dataclass
-class NixlTransportMetadata(TensorTransportMetadata):
-    """Metadata for tensors stored in the GPU object store for NIXL transport.
-
-    Args:
-        nixl_serialized_descs: Serialized tensor descriptors for NIXL transport.
-        nixl_agent_meta: The additional metadata of the remote NIXL agent.
-    """
-
-    nixl_reg_descs: Optional[Any] = None
-    nixl_serialized_descs: Optional[bytes] = None
-    nixl_agent_meta: Optional[bytes] = None
-
-    __eq__ = object.__eq__
-    __hash__ = object.__hash__
-
-
-@dataclass
-class CollectiveTransportMetadata(TensorTransportMetadata):
-    """Metadata for tensors stored in the GPU object store for collective transport."""
-
-
-@dataclass
-class CommunicatorMetadata:
-    """Metadata for the communicator.
-
-    Args:
-        communicator_name: The name of the communicator.
-    """
-
-    communicator_name: str = ""
-
-
-@dataclass
-class CollectiveCommunicatorMetadata(CommunicatorMetadata):
-    """Metadata for the collective communicator (e.g. NCCL, GLOO).
-
-    Args:
-        src_rank: The rank of the source actor.
-        dst_rank: The rank of the destination actor.
-    """
-
-    src_rank: Optional[int32] = None
-    dst_rank: Optional[int32] = None
-
-
-@dataclass
-class NixlCommunicatorMetadata(CommunicatorMetadata):
-    """Metadata for the NIXL communicator."""
-
-
 class ReduceOp(Enum):
     SUM = 0
     PRODUCT = 1
@@ -130,9 +63,6 @@ class ReduceOp(Enum):
 
 
 unset_timeout_ms = timedelta(milliseconds=-1)
-
-# This is used to identify the collective group for NIXL.
-NIXL_GROUP_NAME = "ray_internal_nixl_group"
 
 
 @dataclass
